Split all model classes
