+++ src/TF2HUD.Editor/Models/Option.cs
@@ -1,0 +1,18 @@
+﻿using System.Collections.Generic;
+using System.Text.Json.Serialization;
+using System.Windows;
+using Newtonsoft.Json.Linq;
+
+namespace HUDEditor.Models
+{
+    public class Option
+    {
+        [JsonPropertyName("FileName")] public string FileName;
+        [JsonPropertyName("Files")] public JObject Files;
+        [JsonPropertyName("Label")] public string Label;
+        [JsonPropertyName("RenameFile")] public RenameFile RenameFile;
+        [JsonPropertyName("Special")] public string Special;
+        [JsonPropertyName("SpecialParameters")] public string[] SpecialParameters;
+        [JsonPropertyName("Value")] public string Value;
+    }
+}
